// src/types/article.ts
export interface Article {
    abstract: string
    byline: Byline
    document_type: string
    headline: Headline
    _id: string
    keywords: Keyword[]
    multimedia: Multimedia
    news_desk: string
    print_page: string
    print_section: string
    pub_date: string
    section_name: string
    snippet: string
    source: string
    subsection_name: string
    type_of_material: string
    uri: string
    web_url: string
    word_count: number
}

export interface Byline {
    original: string
}

export interface Headline {
    main: string
    kicker: string
    print_headline: string
}

export interface Keyword {
    name: string
    value: string
    rank: number
}

export interface Multimedia {
    caption: string
    credit: string
    default: Default
    thumbnail: Thumbnail
}

export interface Default {
    url: string
    height: number
    width: number
}

export interface Thumbnail {
    url: string
    height: number
    width: number
}